UFO Sightings
A web page created for educational purposes that contains data on Unidentified Flying Objects. The dataset provided is in the form of an array of JavaScript objects. User can set multiple filters and search for UFO sightings using the following criteria based on the table columns:  date/time, city, state, country, shape.